Fix condition in AsyncBase and add test.
